refactor: use `uint` for object ID-s

+++ Ship_Game/Data/Binary/CollectionState.cs
@@ -7,9 +7,9 @@ namespace Ship_Game.Data.Binary;
 // T[], Array<T>, Map<K,V> or HashSet<T>
 public class CollectionState : ObjectState
 {
-    public int[] Items;
+    public uint[] Items;
 
-    public CollectionState(object obj, int id) : base(obj, id)
+    public CollectionState(object obj, uint id) : base(obj, id)
     {
     }
 
@@ -17,17 +17,17 @@ public class CollectionState : ObjectState
     {
         if (Items == null)
         {
-            w.BW.WriteVLu32((uint)0); // collection length
+            w.BW.WriteVLu32(0u); // collection length
         }
         else
         {
             w.BW.WriteVLu32((uint)Items.Length); // collection length
             for (int i = 0; i < Items.Length; ++i)
-                w.BW.WriteVLu32((uint)Items[i]);
+                w.BW.WriteVLu32(Items[i]);
         }
     }
 
-    public override void Remap(int[] map)
+    public override void Remap(uint[] map)
     {
         Remap(map, Items);
     }
@@ -44,18 +44,18 @@ public class CollectionState : ObjectState
 
         if (coll is MapSerializer maps)
         {
-            Items = new int[count * 2];
+            Items = new uint[count * 2];
             var e = ((IDictionary)Obj).GetEnumerator();
             for (int i = 0; i < count && e.MoveNext(); ++i)
             {
-                int keyId = scanner.ScanObjectState(maps.KeySerializer, e.Key);
+                uint keyId = scanner.ScanObjectState(maps.KeySerializer, e.Key);
                 Items[i*2+0] = keyId;
                 SetValue(scanner, i*2+1, e.Value, valCanBeNull, typeMap);
             }
         }
         else if (coll is HashSetSerializer)
         {
-            Items = new int[count];
+            Items = new uint[count];
             var e = ((IEnumerable)Obj).GetEnumerator();
             for (int i = 0; i < count && e.MoveNext(); ++i)
             {
@@ -64,7 +64,7 @@ public class CollectionState : ObjectState
         }
         else
         {
-            Items = new int[count];
+            Items = new uint[count];
             for (int i = 0; i < count; ++i)
             {
                 object element = coll.GetElementAt(Obj, i);
@@ -83,7 +83,7 @@ public class CollectionState : ObjectState
         {
             // NOTE: VALUES CAN USE ABSTRACT TYPES, SO TYPE CHECK IS REQUIRED FOR EACH ELEMENT
             TypeSerializer item = typeMap.Get(instance!.GetType());
-            int valId = scanner.ScanObjectState(item, instance);
+            uint valId = scanner.ScanObjectState(item, instance);
             Items[index] = valId;
         }
     }
